pyproject.toml
[project]
name = "zero-shot-classifier"
version = "0.1.0"
description = "Zero-shot text classification using a BiEncoder architecture"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "torch>=2.0",
    "transformers>=4.40",
    "huggingface-hub>=0.20",
    "pyyaml>=6.0",
    "google-genai>=1.0",
    "typer>=0.9",
    "tqdm>=4.60",
    "datasets>=2.0",
    "python-dotenv>=1.0",
    "gradio>=6.8.0",
    "accelerate>=1.12.0",
]

[project.optional-dependencies]
dev = ["ipython", "jupyter"]
